+++ Sudoku.Z3Solver/Z3IntSolverSimple.cs
@@ -9,10 +9,8 @@ using System.Diagnostics.CodeAnalysis;
 
 namespace Sudoku.Z3Solver
 {
-    // Exemple de réutilisation en utilisant l'API de scope
-    public class Z3SolverReusableScope : Z3SolverBase
+    public class Z3IntSolverSimple : Z3IntSolverBase
     {
-
         public override SudokuGrid Solve(SudokuGrid s)
         {
             SudokuGrid solution = new SudokuGrid();
@@ -24,17 +22,21 @@ namespace Sudoku.Z3Solver
 
         public void SudokuSolve(SudokuGrid grid, ref SudokuGrid solution)
         {
-            var solver = ReusableSolver;
-            solver.Push();
-			BoolExpr instance_c = GetPuzzleConstraints(grid);
-            solver.Assert(instance_c);
-			if (solver.Check(instance_c) == Status.SATISFIABLE)
-			{
-				Model m = solver.Model;
+            var sudoku_c = GenericContraints;
+            var instance_c = GetPuzzleConstraints(grid);
+            Solver s = ctx.MkSolver();
+            s.Assert(sudoku_c);
+            s.Assert(instance_c);
+
+            if (s.Check() == Status.SATISFIABLE)
+            {
+                Model m = s.Model;
                 for (uint i = 0; i < 9; i++)
-                {   
+                {
                     for (uint j = 0; j < 9; j++)
-					    solution.Cells[i][j] = ((IntNum)m.Evaluate(X[i][j])).Int;
+                    {
+                        solution.Cells[i][j] = ((IntNum)m.Evaluate(CellVariables[i][j])).Int;
+                    }
                 }
             }
             else
@@ -42,8 +44,6 @@ namespace Sudoku.Z3Solver
                 Console.WriteLine("Failed to solve sudoku");
                 throw new Exception("Failed to solve sudoku");
             }
-			solver.Pop();
-
         }
-    }   
+    }
 }
